CHE-84 Forum controller directory restructure
