+++ server/controllers/forumController/index.ts
@@ -1,0 +1,7 @@
+import addForum from './addForum/addForum';
+import getAllForums from './getAllForums/getAllForums';
+import getForumById from './getForumById/getForumById';
+import updateForum from './updateForum/updateForum';
+import deleteForum from './deleteForum/deleteForum';
+
+export { addForum, getAllForums, getForumById, updateForum, deleteForum };
